+++ mock/data/motebehov.json
@@ -1,1 +1,50 @@
-[]
+[
+  {
+    "UUID": "198a6dbf-c987-4b57-a401-a3915ec11429198a6dbf-c987-4b57-a401-a3915ec11429",
+    "id": "b415b37c-ee10-44b6-bddf-54d049ef25f9",
+    "opprettetDato": "2019-01-07T13:53:57.047+01:00",
+    "aktorId": "1",
+    "opprettetAv": "1",
+    "virksomhetsnummer": "110110110",
+    "motebehovSvar": {
+      "harMotebehov": true,
+      "friskmeldingForventning": "Nei",
+      "tiltak": "",
+      "tiltakResultat": "",
+      "forklaring": ""
+    },
+    "tildeltEnhet": "0330"
+  },
+  {
+    "UUID": "198a6dbf-c987-4b57-a401-a3915ec11421",
+    "id": "c815ae0b-9e9b-40b0-bd1c-d1c39dc5f481",
+    "opprettetDato": "2019-01-07T13:53:57.047+01:00",
+    "aktorId": "1",
+    "opprettetAv": "2",
+    "virksomhetsnummer": "110110110",
+    "motebehovSvar": {
+      "harMotebehov": false,
+      "friskmeldingForventning": "Nei",
+      "tiltak": "",
+      "tiltakResultat": "",
+      "forklaring": "Jeg liker ikke møte!!"
+    },
+    "tildeltEnhet": "0330"
+  },
+  {
+    "UUID": "198a6dbf-c987-4b57-a401-a3915ec11411",
+    "id": "b415b37c-ee10-44b6-bddf-54d049ef25f2",
+    "opprettetDato": "2019-01-08T13:53:57.047+01:00",
+    "aktorId": "1",
+    "opprettetAv": "1",
+    "virksomhetsnummer": "000999000",
+    "motebehovSvar": {
+      "harMotebehov": true,
+      "friskmeldingForventning": "Nei",
+      "tiltak": "",
+      "tiltakResultat": "",
+      "forklaring": "Møter er bra!"
+    },
+    "tildeltEnhet": "0330"
+  }
+]
