+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "test": "mocha --require babel-core/register --require babel-polyfill --require ignore-styles  ./test/setup.js \"{js/**/*.test,test/**/*}.@(js|jsx)\"",
     "test:watch": "npm run test -- --watch",
-    "lint": "eslint src/js/. --ext .js",
+    "lint": "eslint src/. --ext .js",
     "lint:fix": "npm-run-all lintJs:fix lintTest:fix",
     "lintJs": "eslint js/. --ext .js",
     "lintJs:fix": "eslint js/. --ext .js --fix",
