--- src/sagas/util.js
@@ -1,9 +1,0 @@
-export const finnMiljoStreng = () => {
-    const host = window.location.host;
-    const bindestrekIndex = host.indexOf('-');
-    if (bindestrekIndex === -1) {
-        return '';
-    }
-    const dotIndex = host.indexOf('.');
-    return host.substring(bindestrekIndex, dotIndex);
-};
