+++ src/utils/miljoUtil.js
@@ -1,37 +1,37 @@
-export const finnMiljoStreng = () => {
-  return erPreProd()
-    ? '-q1'
-    : '';
-};
-
 export const erProd = () => {
     return window.location.href.indexOf('nais.adeo.no') > -1;
 };
 
 export const erPreProd = () => {
-  return window.location.href.indexOf('nais.preprod.local') > -1;
+    return window.location.href.indexOf('nais.preprod.local') > -1;
+};
+
+export const finnMiljoStreng = () => {
+    return erPreProd()
+        ? '-q1'
+        : '';
 };
 
 export const erLokal = () => {
-  return window.location.host.indexOf('localhost') > -1;
+    return window.location.host.indexOf('localhost') > -1;
 };
 
 export const finnNaisUrlDefault = () => {
-  return erPreProd() ?
-      '.nais.preprod.local'
-      : '.nais.adeo.no';
+    return erPreProd() ?
+        '.nais.preprod.local'
+        : '.nais.adeo.no';
 };
 
 export const fullNaisUrlDefault = (host, path) => {
-  if (erLokal()) {
-    return path;
-  }
-  return `https://${host}${finnNaisUrlDefault()}${path}`;
+    if (erLokal()) {
+        return path;
+    }
+    return `https://${host}${finnNaisUrlDefault()}${path}`;
 };
 
 export const fullAppAdeoUrl = (path) => {
-  if (erLokal()) {
-    return path;
-  }
-  return `https://app${finnMiljoStreng()}.adeo.no${path}`;
+    if (erLokal()) {
+        return path;
+    }
+    return `https://app${finnMiljoStreng()}.adeo.no${path}`;
 };
